Feat: restructure likes feature components and implement LikesContainer
