+++ src/app/likes/page.tsx
@@ -1,4 +1,4 @@
-import LikesContainer from '@/features/likes/container/LikesContainer';
+import LikesContainer from '@/features/likes/LikesContainer';
 
 export default function LikesPage() {
   return <LikesContainer />;
